# Perf: optimze Queue data structure