--- a/src/Fantomas.Tests/QueueTests.fs
+++ b/src/Fantomas.Tests/QueueTests.fs
@@ -5,12 +5,79 @@ open NUnit.Framework
 open FsUnit
 open FsCheck
 
+module Queue =
+    let ofLists xss = (Queue.empty, xss) ||> List.fold (fun q xs -> Queue.append q xs) 
+
 [<Test>]
 let ``Queue.append``() =
     Check.One (Config.Default, fun xs ys ->
         let q = Queue.ofList xs
         let result q = Queue.append q ys |> Queue.toSeq |> Seq.toList
-        let expected q = (q, ys) ||> List.fold (fun q y -> Queue.conj y q) |> Queue.toSeq |> Seq.toList
-        result q |> should equivalent (expected q)
-        result (Queue.rev q) |> should equivalent (expected (Queue.rev q))
+        let expected x = x @ ys
+        result q |> should equivalent (expected xs)
+        result (Queue.rev q |> Queue.ofSeq) |> should equivalent (expected (List.rev xs))
+        )
+
+// [<Test>]
+// let ``Queue.conj``() =
+//     Check.One (Config.Default, fun xs ys ->
+//         let result = Queue.ofList (xs @ ys)
+//         let expected = (Queue.ofList xs, ys) ||> List.fold (fun q y -> Queue.conj y q) |> Queue.toSeq |> Seq.toList
+//         result |> should equivalent expected
+//         )
+
+[<Test>]
+let ``Queue.tryHead``() =
+    Check.One (Config.Default, fun xs ->
+        let result = Queue.ofList xs |> Queue.tryHead
+        let expected = xs |> List.tryHead
+        result |> should equal expected
+        )
+
+[<Test>]
+let ``Queue.length``() =
+    Check.One (Config.Default, fun xs ->
+        let result = Queue.ofList xs |> Queue.length
+        let expected = xs |> List.length
+        result |> should equal expected
+        )
+
+[<Test>]
+let ``Queue.rev``() =
+    Check.One (Config.Default, fun xs ->
+        let result = Queue.ofList xs |> Queue.rev |> Seq.toList
+        let expected = xs |> List.rev
+        result |> should equivalent expected
+        )
+
+[<Test>]
+let ``Queue.toSeq``() =
+    Check.One (Config.Default, fun xs ->
+        let result = Queue.ofList xs |> Queue.toSeq |> Seq.toList
+        let expected = xs
+        result |> should equivalent expected
+        )
+
+[<Test>]
+let ``Queue.skipExists``() =
+    Check.One (Config.Default, fun xss n f ->
+        n <= List.sumBy List.length xss ==>
+        let result = Queue.ofLists xss |> Queue.skipExists n f
+        let expected = xss |> List.collect id |> Seq.skip n |> Seq.exists f
+        result |> should equal expected
+        )
+
+
+[<Test>]
+let ``Queue ref transp``() =
+    Check.One (Config.Default, fun xs ys zs ->
+        let result1 = Queue.ofList xs
+        let result2 = Queue.append result1 ys 
+        let result3 = Queue.append result1 zs 
+        let expected1 = xs
+        let expected2 = xs @ ys
+        let expected3 = xs @ zs
+        result1 |> should equivalent expected1
+        result2 |> should equivalent expected2
+        result3 |> should equivalent expected3
         )