This is total rewrite of `Queue` data structure for performance for our use case.

before 

| Method |    Mean |   Error |  StdDev | Rank |        Gen 0 |       Gen 1 |      Gen 2 | Allocated |
|------- |--------:|--------:|--------:|-----:|-------------:|------------:|-----------:|----------:|
| Format | 35.68 s | 6.544 s | 0.359 s |    1 | 2376000.0000 | 777000.0000 | 82000.0000 |  11.92 GB |

after

| Method |    Mean |   Error |  StdDev | Rank |        Gen 0 |       Gen 1 |      Gen 2 | Allocated |
|------- |--------:|--------:|--------:|-----:|-------------:|------------:|-----------:|----------:|
| Format | 33.09 s | 8.043 s | 0.441 s |    1 | 2165000.0000 | 668000.0000 | 79000.0000 |  10.75 GB |